Fix splash screen, gave radius to covers, fix listitem padding
